requirements.txt
argh==0.26.2
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
backpack==0.1
bcrypt==3.1.4
blinker==1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cleo==0.6.8
cryptography==2.3
Cython==0.28.5
Faker==0.7.18
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
masonite==2.0.22
masonite-dashboard==0.1.3
masonite-entry==0.1.6
masonite-scheduler==1.0.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
orator==0.9.7
passlib==1.7.1
pastel==0.1.0
pathtools==0.1.2
pendulum==1.5.1
pluggy==0.6.0
psutil==5.4.6
py==1.6.0
pyaml==16.12.2
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
pylev==1.3.0
pylint==2.1.1
PyMySQL==0.9.2
pytest==3.6.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.5
pytzdata==2018.5
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
simplejson==3.16.0
six==1.11.0
slugify==0.0.1
tabulate==0.8.2
tldextract==2.2.0
typed-ast==1.1.0
tzlocal==1.5.1
unicode==2.6
Unidecode==1.0.22
urllib3==1.23
validator.py==1.2.5
waitress==1.1.0
watchdog==0.8.3
whitenoise==3.3.1
wrapt==1.10.11
